requirements.txt
flask==2.0.1
pyrebase4==4.5.0
python-dotenv==0.19.0 